backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
PyJWT>=2.8.0
python-multipart>=0.0.9
deep-translator>=1.11.4
requests>=2.31.0
